removed methods that do not leverage the DB from the store and moved them to utils.js. Eventually utils.js should be broken down
