+++ authentication/store.js
@@ -1,69 +1,12 @@
 /**
  * http://usejsdoc.org/
  */
-const crypto = require('crypto')
 const knex = require('knex')(require('../knexfile'))
 const utils = require('../utils/utils')
+const logger = require('../logger')
 const constants = require('../constants')
 
 // BEGIN UTILITY FUNCTIONS
-
-function validatePassword(password){
-	/*
-	 * OWASP defines a strong password as:
-	 * 
-	 * 1) Password Length Minimum length should be enforced by the application
-	 * Passwords shorter than 10 characters are considered weak Typical maximum
-	 * password length is 128 characters Passphrases shorter than 20 characters
-	 * are usually considered weak if they only consist of lower case latin
-	 * characters 2) Password Complexity The application should enforce password
-	 * complexity rules to discourage easy passwords Password mechanisms should
-	 * allow virtually any character the user can type including spaces
-	 * Passwords should be case sensitive An example of basic complexity
-	 * checking would be: password must contain 3/4 of the following rules: at
-	 * least 1 uppercase character (A-Z) at least 1 lowercase character (a-z) at
-	 * least 1 digit (0-9) at least 1 special character at least 10 characters
-	 * at most 128 characters not more than 2 identicaly characters in a row
-	 * (e.g., 111 not allowed)
-	 */
-	
-	/*
-	 * This regex enforces these rules: 1) At least one uppercase english letter
-	 * [A-Z] 2) At least one lowercase english letter [a-z] 3) At least one
-	 * digit [0-9] 4) At least one special character [#?!@$% ^&*-] 5) Minimum
-	 * length 10, maximum length 128
-	 */
-	let regex = /^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$ %^&*-]).{10,128}$/
-	let result = regex.test(password)
-	if(!result){
-		// build error message
-		let message = "Your password does not meet the following requirement(s):"
-		let upperCaseRegex = /^.*[A-Z].*$/
-		let lowerCaseRegex = /^.*[a-z].*$/ 
-		let digitRegex = /^.*[0-9].*$/ 
-		let specialCharRegex = /^.*[#?!@$% ^&*-].*$/ 
-	    if(!upperCaseRegex.test(password)){
-	    	message += ' At least one upper case english letter [A-Z].'
-	    }
-		if(!lowerCaseRegex.test(password)){
-			message += ' At least one lower case english letter [a-z].'
-		}
-		if(!digitRegex.test(password)){
-			message += ' At least one digit [0-9].'
-		}
-		if(!specialCharRegex.test(password)){
-			message += ' At lease one special character [#?!@$ %^&*-].'
-		}
-		if(password.length < 10 || password.length > 128){
-			message += ' A password length between 10 and 128 characters.'
-		}
-		console.log(message)
-		return {success: false, message: message}
-	}
-	else{
-		return {success: true}
-	}
-}
 
 // want to prevent brute force login attempts
 /*
@@ -96,34 +39,14 @@ function incrementLockCount(user){
 	// Unlock their account after 1 minute
 	if(user.lock_count === 2){
 		setTimeout(function(){
-			console.log('unlocking user', user)
+			logger.info('Unlocking user with id ::' + user.id)
 			resetLockCount(user).then(()=>{
-				console.log('user unlocked')
+				logger.info('User ID ::' + user.id + ':: unlocked')
 			})
 		}, 60*1000)
 	}
 	return knex('user').increment('lock_count').where({id: user.id})
 }
-
-/**
- * we want to store encrypted passwords in the DB Accept a salt and only
- * generate one if none is supplied
- * 
- * @param {password:MISSING,salt:utils.randomString()}
- * @returns
- */
-function saltHashPassword ({
-	  password,
-	  salt = utils.randomString()
-	}) {
-	  const hash = crypto
-	    .createHmac('sha512', salt)
-	    .update(password)
-	  return {
-	    salt,
-	    hash: hash.digest('hex')
-	  }
-	}
 
 function authenticate ({ username, password }) {
 	username = username.toLowerCase() // username should be lower case in the
@@ -209,7 +132,6 @@ function getUserID(username) {
   }
 
 module.exports = {
-		validatePassword,
 		saltHashPassword,
 		createUser,
 		createAdvertiser,
